// agency/lib/seo/types.ts
export interface SeoPostalAddress {
  streetAddress?: string
  addressLocality?: string
  addressRegion?: string
  postalCode?: string
  addressCountry?: string
}

export interface SeoContactPoint {
  contactType: string
  telephone?: string
  email?: string
  areaServed?: string | string[]
  availableLanguage?: string | string[]
}

export interface SeoBrandConfig {
  brandName: string
  brandUrl: string
  brandLogo: string
  brandDescription: string
  brandSameAs: readonly string[]
  brandTelephone?: string
  brandEmail?: string
  brandLocation?: SeoPostalAddress
  brandServiceArea: readonly string[]
  brandContactPoints: readonly SeoContactPoint[]
}

export interface SeoSiteConfig {
  siteUrl: string
  defaultLocale: string
  locales: readonly string[]
  titleTemplate: string
  defaultTitle: string
  defaultDescription: string
}

export interface SeoConfig {
  brand: SeoBrandConfig
  site: SeoSiteConfig
}

export type SeoRobots = Metadata["robots"]
export type SeoAlternates = Metadata["alternates"]

export type SeoOpenGraphImage = {
  url: string
  width?: number
  height?: number
  alt?: string
}

export type SeoTwitterImage = string | SeoOpenGraphImage

export interface SeoPageMetadataInput {
  title: string
  description: string
  canonicalPath?: string
  robots?: SeoRobots
  openGraph?: Omit<NonNullable<Metadata["openGraph"]>, "title" | "description" | "url" | "siteName" | "images"> & {
    type?: string
    images?: SeoOpenGraphImage[]
  }
  twitter?: Omit<NonNullable<Metadata["twitter"]>, "title" | "description" | "images"> & {
    card?: string
    images?: SeoTwitterImage[]
  }
  alternates?: SeoAlternates
}

export interface SeoLayoutMetadataInput {
  defaultTitle?: string
  description?: string
  canonicalPath?: string
  robots?: SeoRobots
}